+++ CHANGELOG.md
@@ -183,6 +183,10 @@ instances according to the [policy configuration](https://github.com/nasa/cumulu
 - **CUMULUS-2684**
   - Added optional collection level parameter `s3MultipartChunksizeMb` to collection's `meta` field
   - Updated `move-granules` task to take in an optional config parameter s3MultipartChunksizeMb
+- **CUMULUS-2703**
+  - Updated reconciliation report lambda to use createdAt for report time range instead of updatedAt
+  - Added `@cumulus/api/lambdas/reports/orca-backup-reconciliation-report` to create
+    `ORCA Backup` reconciliation report
 - **CUMULUS-2747**
   - Updated data management type doc to include additional fields for provider configurations
 - **CUMULUS-2773**